test running project through fe

+++ backend/src/main/java/com/example/backend/controller/CategoryController.java
@@ -24,7 +24,7 @@ public class CategoryController {
     @GetMapping
     public ResponseEntity<ApiResponse<List<CategoryResponse>>> getCategories(
             @RequestParam(required = false) String type) {
-        Long userId = getCurrentUserId();
+        Integer userId = getCurrentUserId();
         List<CategoryResponse> categories = categoryService.getCategories(userId, type);
         return ResponseEntity.ok(ApiResponse.success("Lấy danh sách danh mục thành công", categories));
     }
@@ -32,7 +32,7 @@ public class CategoryController {
     @PostMapping
     public ResponseEntity<ApiResponse<CategoryResponse>> createCategory(
             @Valid @RequestBody CategoryRequest request) {
-        Long userId = getCurrentUserId();
+        Integer userId = getCurrentUserId();
         CategoryResponse response = categoryService.createCategory(userId, request);
         return ResponseEntity.status(HttpStatus.CREATED)
                 .body(ApiResponse.success("Tạo danh mục thành công", response));
@@ -40,23 +40,23 @@ public class CategoryController {
 
     @PutMapping("/{id}")
     public ResponseEntity<ApiResponse<CategoryResponse>> updateCategory(
-            @PathVariable Long id,
+            @PathVariable Integer id,
             @Valid @RequestBody CategoryRequest request) {
-        Long userId = getCurrentUserId();
+        Integer userId = getCurrentUserId();
         CategoryResponse response = categoryService.updateCategory(userId, id, request);
         return ResponseEntity.ok(ApiResponse.success("Cập nhật danh mục thành công", response));
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<ApiResponse<Map<String, String>>> deleteCategory(@PathVariable Long id) {
-        Long userId = getCurrentUserId();
+    public ResponseEntity<ApiResponse<Map<String, String>>> deleteCategory(@PathVariable Integer id) {
+        Integer userId = getCurrentUserId();
         categoryService.deleteCategory(userId, id);
         return ResponseEntity.ok(ApiResponse.success("Xóa danh mục thành công",
                 Map.of("message", "Xóa danh mục thành công")));
     }
 
-    private Long getCurrentUserId() {
+    private Integer getCurrentUserId() {
         String name = SecurityContextHolder.getContext().getAuthentication().getName();
-        return Long.parseLong(name);
+        return Integer.parseInt(name);
     }
 }
